Fixes and Upgrades
1) fixed tables displayed during sampling to properly show fields of included documents
2) Change LDA to use multicore approach to improve model training speed
3) added gate keeping into multiprocessing to reduce change of LDA muticore + a multiprocessing function overwhelming application
4) resized chord plots to better fit the plan allocated to them.
5) reworked collection panel so that data details show as two rows rather then as a single column

+++ src/Coding/ModuleCoding.py
@@ -140,7 +140,7 @@ class CodingDatasetPanel(wx.Panel):
             if isinstance(node, Datasets.Document):
                 self.Freeze()
                 if node.key not in self.document_windows:
-                    self.document_windows[node.key] = DatasetGUIs.DocumentPanel(self.splitter, node, size=(self.GetSize().GetWidth(), self.GetSize().GetHeight()/4*3))
+                    self.document_windows[node.key] = DatasetGUIs.DocumentPanel(self.splitter, node, size=(self.GetSize().GetWidth(), int(self.GetSize().GetHeight()/4*3)))
                 
                 bottom_window = self.splitter.GetWindow2()
                 bottom_window.Hide()
